feat: enhance Archive command with album support
Add comprehensive album support to archive operations:
- Album models: AlbumConfiguration, AlbumEntity, AlbumDateRange, AlbumReverseGeocode
- Album types and history tracking
- ArchiveDatabaseOptions for database-specific configuration
- Enhanced ArchiveRunner with album creation and management
- DbService methods for album operations (create, query, update)
- MissingReverseGeocodeAction enum for handling missing location data
- ActionableReverseGeocodeValidator for validating geocode operations
- Comprehensive database integration tests for album functionality
- Test fixtures with Unix and Windows archive structures
- End-to-end and unit tests for archive with albums
- Add DateTimeExtensions utility for date/time operations
- Add ValidEnumValidator for enum validation
- Update model classes with minor improvements (Coordinate, ExifData, PhotoEntity)
- Enhance validators with better error messages and validation logic
- Update options classes with improved property handling
- Update utility classes (Constants, StringExtensions, ValidatorExtensions)
- Enhance FileExistsOnDestinationPathException with better context
- Update all test fakes with new fields and methods
- Update end-to-end tests for all commands
- Update FluentValidation tests with new test cases
- Update test utilities and helpers

+++ tests/Fakes/DateTimeFakes.cs
@@ -9,22 +9,22 @@ public static class DateTimeFakes
 	private const int HourDefault = 0;
 	private const int SecondDefault = 0;
 
-	public static DateTime WithDayHour(int day, int hour)
+	public static DateTime WithDayHour(byte day, byte hour)
 	{
 		return new DateTime(YearDefault, MonthDefault, day, hour, MinuteDefault, SecondDefault);
 	}
 
-	public static DateTime WithMinuteSecond(int minute, int second)
+	public static DateTime WithMinuteSecond(byte minute, byte second)
 	{
 		return new DateTime(YearDefault, MonthDefault, DayDefault, HourDefault, minute, second);
 	}
 
-	public static DateTime WithDay(int day)
+	public static DateTime WithDay(byte day)
 	{
 		return new DateTime(YearDefault, MonthDefault, day);
 	}
 
-	public static DateTime WithMonth(int month)
+	public static DateTime WithMonth(byte month)
 	{
 		return new DateTime(YearDefault, month, DayDefault);
 	}
@@ -34,25 +34,25 @@ public static class DateTimeFakes
 		return new DateTime(year, MonthDefault, DayDefault);
 	}
 
-	public static string FormatDay(int day)
+	public static string FormatDay(byte day)
 	{
 		var dateTime = WithDay(day);
 		return dateTime.ToString(ToolOptionFakes.DateFormatWithDay);
 	}
 
-	public static string DirectoryFormatDay(int day)
+	public static string DirectoryFormatDay(byte day)
 	{
 		var dateTime = WithDay(day);
 		return Path.Combine(dateTime.ToString(ToolOptionFakes.YearFormat), dateTime.ToString(ToolOptionFakes.MonthFormat), dateTime.ToString(ToolOptionFakes.DayFormat));
 	}
 
-	public static string FormatMonth(int month)
+	public static string FormatMonth(byte month)
 	{
 		var dateTime = WithMonth(month);
 		return dateTime.ToString(ToolOptionFakes.DateFormatWithMonth);
 	}
 
-	public static string DirectoryFormatMonth(int month)
+	public static string DirectoryFormatMonth(byte month)
 	{
 		var dateTime = WithMonth(month);
 		return Path.Combine(dateTime.ToString(ToolOptionFakes.YearFormat), dateTime.ToString(ToolOptionFakes.MonthFormat));
@@ -75,7 +75,7 @@ public static class DateTimeFakes
 		return dateTime.ToString(ToolOptionFakes.DateTimeFormatWithMinutes);
 	}
 
-	public static DateTime WithSecond(int second)
+	public static DateTime WithSecond(byte second)
 	{
 		return new DateTime(YearDefault, MonthDefault, DayDefault, HourDefault, MinuteDefault, second);
 	}
@@ -85,7 +85,7 @@ public static class DateTimeFakes
 		return DateTime.MinValue;
 	}
 
-	public static string FormatSecond(int second)
+	public static string FormatSecond(byte second)
 	{
 		var dateTime = WithSecond(second);
 		return dateTime.ToString(ToolOptionFakes.DateTimeFormatWithSecondsDefault);
